started working on testing

+++ labs/inheritance/src/main/java/inheritance/Restaurant.java
@@ -14,7 +14,7 @@ public class Restaurant extends Business {
     public void setPrice(String price) { this.price = price; }
 
     public String toString() {
-        return String.format("%s has a rating of %d stars and is in the %s price category.", this.getName(), this.getStars(), this.price);
+        return String.format("%s has a rating of %d stars and is in the %s price category.", this.getName(), (int)this.getStars(), this.price);
     }
 
 }
